+++ R/ids.R
@@ -32,8 +32,11 @@ swc2segmentid <- function(x, include.fragment=FALSE) {
 #' @rdname fafbseg-ids
 #' @examples
 #' \donttest{
-#' segmentid2zip(10001654273)
-#' segmentid2zip(swc2segmentid("10001654273.1.swc"))
+#' # NB the default segmentation for fafbseg (flywire) no longer implies a local
+#' # collection of skeletons, wrap calls in with_segmentation
+#' with_segmentation("20190805", segmentid2zip(10001654273))
+#' with_segmentation("20190805",
+#'   segmentid2zip(swc2segmentid("10001654273.1.swc")))
 #' }
 segmentid2zip <- function(x) {
   divisor <- find_zip_divisor(getOption("fafbseg.skelziproot"))
@@ -152,9 +155,10 @@ ngl_layers <- function(x) {
 }
 
 null2na <- function(x) sapply(x, function(y) if(is.null(y)) NA else y,USE.NAMES = F)
-ngl_segmentation <- function(x=getOption('fafbseg.sampleurl')) {
+ngl_segmentation <- function(x=getOption('fafbseg.sampleurl'), rval=c('url', 'full')) {
+  rval=match.arg(rval)
   layers=ngl_layers(x)
-  sources=sapply(layers, "[[", "source")
+  sources=sapply(layers, function(x) unlist(x[['source']],use.names = F)[1])
   types=sapply(layers, "[[", "type")
   st = data.frame(
     source = null2na(sources),
@@ -165,5 +169,11 @@ ngl_segmentation <- function(x=getOption('fafbseg.sampleurl')) {
   # remove any layers without defined sources
   st=st[!is.na(st$source),,drop=FALSE]
   seglayer=grep('seg', st$type)
-  if(length(seglayer)) layers[[seglayer[1]]] else NULL
+  if(length(seglayer)==0) {
+    NULL
+  } else if(rval=='url') {
+    st$source[[seglayer[1]]]
+  } else {
+    layers[[seglayer[1]]]
+  }
 }