+++ tests/testthat/test-ids.R
@@ -61,6 +61,6 @@ test_that('neuroglancer segmentation for all built-in urls',  {
   releases=eval(formals(choose_segmentation)[['release']])
   for(r in releases) {
     with_segmentation(r,
-      expect_is(ngl_segmentation()$source, 'character'))
+      expect_is(ngl_segmentation(rval='url'), 'character'))
     }
 })